+++ src/init.cpp
@@ -10,7 +10,7 @@ extern "C" {
     {NULL, NULL, 0}
   };
 
-  void R_init_toth(DllInfo *dll) {
+  void R_init_thoth(DllInfo *dll) {
     R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
     R_useDynamicSymbols(dll, FALSE);
   }