+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(thoth)
+library(testthat)
+
+test_check("thoth")